Better error handling/reporting for the deployment tool

+++ Convert-Shortcut.ps1
@@ -12,6 +12,8 @@ function Write-LogString {
 # Get all .website shortcuts on the C: drive of the device
 Write-LogString 'Getting shortcuts on this device...'
 $Shortcuts = Get-ChildItem -Path C:\ -Recurse -Filter *.website -ErrorAction SilentlyContinue
+# Variable to return on script completion
+$ReturnCode = 0
 # Check if there are shortcuts to update
 if ($Shortcuts.Count -gt 0) {
 	Write-LogString "Processing $($Shortcuts.Count) shortcut$(if($Shortcuts.Count -gt 1){'s'})..."
@@ -21,36 +23,47 @@ if ($Shortcuts.Count -gt 0) {
 		Write-LogString 'Getting file name for shortcut...'
 		$ShortcutName = [System.IO.Path]::GetFileNameWithoutExtension($Shortcut)
 		Write-LogString $ShortcutName
-		# Get the URL from the old shortcut
-		Write-LogString 'Getting URL from old shortcut...'
-		$ShortcutURL = Get-Content $Shortcut.FullName | ForEach-Object { if ($_.StartsWith('URL=')) { return $_.Split('=')[1] } }
-		Write-LogString $ShortcutURL
-		# Generate a new location for the shortcut
-		Write-LogString 'Generate new shortcut location...'
-		$NewShortcutLocation = "$($Shortcut.DirectoryName)\$($ShortcutName).url"
-		Write-LogString $NewShortcutLocation
-		# Check if the shortcut already exists
-		if (Test-Path $NewShortcutLocation) {
-			# Shortcut already exists - stop and return error code to deployment tool
-			Write-LogString "Shortcut already exists at $($NewShortcutLocation)"
-			Write-LogString 'Halting script to allow for manual intervention'
-			exit 1
-		} else {
-			# Create a shell object
-			Write-LogString 'Creating shell object...'
-			$Shell = New-Object -ComObject ('WScript.Shell')
-			# Create the new shortcut in the same directory as the old one
-			Write-LogString 'Creating shortcut...'
-			$NewShortcut = $Shell.CreateShortcut($NewShortcutLocation)
-			# Set the URL to the one we got earlier
-			Write-LogString 'Setting target path to correct URL...'
-			$NewShortcut.TargetPath = $ShortcutURL
-			# Save the shortcut
-			Write-LogString 'Saving changes to the shortcut...'
-			$NewShortcut.Save()
-			# Remove the old shortcut (to avoid confusion for users)
-			Write-LogString 'Remove the old shortcut...'
-			Remove-Item $Shortcut.FullName -Force
+		try {
+			# Get the URL from the old shortcut
+			Write-LogString 'Getting URL from old shortcut...'
+			$ShortcutURL = Get-Content $Shortcut.FullName | ForEach-Object { if ($_.StartsWith('URL=')) { return $_.Split('=')[1] } }
+			Write-LogString $ShortcutURL
+		} catch {
+			# If unable to get the URL for the shortcut - log issue and set variable to false
+			Write-LogString "Unable to get the URL for $($ShortcutName)!"
+			$ShortcutURL = $false
+			# Bump return code so the deployment tool shows an error
+			$ReturnCode++
+		}
+		# Provided a shortcut URL is found
+		if ($ShortcutURL) {
+			# Generate a new location for the shortcut
+			Write-LogString 'Generate new shortcut location...'
+			$NewShortcutLocation = "$($Shortcut.DirectoryName)\$($ShortcutName).url"
+			Write-LogString $NewShortcutLocation
+			# Check if the shortcut already exists
+			if (Test-Path $NewShortcutLocation) {
+				# Shortcut already exists - log issue and continue
+				Write-LogString "Shortcut already exists at $($NewShortcutLocation)"
+				# Bump return code so the deployment tool shows an error
+				$ReturnCode++
+			} else {
+				# Create a shell object
+				Write-LogString 'Creating shell object...'
+				$Shell = New-Object -ComObject ('WScript.Shell')
+				# Create the new shortcut in the same directory as the old one
+				Write-LogString 'Creating shortcut...'
+				$NewShortcut = $Shell.CreateShortcut($NewShortcutLocation)
+				# Set the URL to the one we got earlier
+				Write-LogString 'Setting target path to correct URL...'
+				$NewShortcut.TargetPath = $ShortcutURL
+				# Save the shortcut
+				Write-LogString 'Saving changes to the shortcut...'
+				$NewShortcut.Save()
+				# Remove the old shortcut (to avoid confusion for users)
+				Write-LogString 'Remove the old shortcut...'
+				Remove-Item $Shortcut.FullName -Force
+			}
 		}
 	}
 }
@@ -60,4 +73,4 @@ else {
 }
 # End of script - exit with code 0
 Write-LogString 'Script has finished'
-exit 0
+exit $ReturnCode
